requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Security-Too==5.7.1
Flask-WTF==1.2.2
email-validator==2.3.0
argon2-cffi==25.1.0
python-dotenv==1.2.2
pytest==9.0.2
